Move actions into separate folder
