+++ traffic_editor/gui/actions/add_vertex.cpp
