+++ traffic_editor/gui/editor.cpp
@@ -41,7 +41,7 @@
 #include "ament_index_cpp/get_package_prefix.hpp"
 #include "ament_index_cpp/get_resource.hpp"
 
-#include "add_vertex.h"
+#include "actions/add_vertex.h"
 #include "add_param_dialog.h"
 #include "building_dialog.h"
 #include "building_level_dialog.h"
